+++ lib/view/user/user_inof.dart
@@ -1,5 +1,5 @@
-import 'package:employe_portal/view/login_screen.dart';
-import 'package:employe_portal/view/pdfviewer.dart';
+import 'package:employe_portal/view/admin/login_screen.dart';
+import 'package:employe_portal/view/admin/pdfviewer.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:google_fonts/google_fonts.dart';
@@ -24,29 +24,58 @@ class _User_InfoState extends State<User_Info> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        leading: IconButton(
-          icon: const Icon(
-            Icons.logout,
-            color: Color(0xff2476BD),
-            size: 25,
-          ),
-          onPressed: () async {
-            SharedPreferences prefs = await SharedPreferences.getInstance();
-            await prefs.clear(); // Clear all SharedPreferences
 
-            Navigator.pushAndRemoveUntil(
-              context,
-              MaterialPageRoute(builder: (context) => Login_Screen()),
-              (Route<dynamic> route) =>
-                  false, // Prevent going back to revious screen
-            );
-          },
-        ),
+        leading: IconButton(
+  icon: const Icon(
+    Icons.logout,
+    color: Color(0xff2476BD),
+    size: 25,
+  ),
+  onPressed: () async {
+    // Show a confirmation dialog
+    bool confirmLogout = await showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text('Logout', style: TextStyle(color:  Color(0xff2476BD, ), fontWeight: FontWeight.bold),),
+          content: const Text('Are you sure you want to logout?', style: TextStyle(color:  Colors.black, fontWeight: FontWeight.bold)),
+          actions: <Widget>[
+            TextButton(
+              child: const Text('Cancel', style: TextStyle(color:  Color(0xff2476BD),  fontWeight: FontWeight.bold), ),
+              onPressed: () {
+                Navigator.of(context).pop(false); // Return false when cancel is pressed
+              },
+            ),
+            TextButton(
+              child: const  Text('Logout',style: TextStyle(color:  Color(0xff2476BD),  fontWeight: FontWeight.bold),),
+              onPressed: () {
+                Navigator.of(context).pop(true); // Return true when logout is confirmed
+              },
+            ),
+          ],
+        );
+      },
+    );
+
+    // If user confirms logout, clear SharedPreferences and navigate to login screen
+    if (confirmLogout ?? false) {
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+      await prefs.clear(); // Clear all SharedPreferences
+
+      Navigator.pushAndRemoveUntil(
+        context,
+        MaterialPageRoute(builder: (context) => Login_Screen()),
+        (Route<dynamic> route) => false, // Prevent going back to previous screen
+      );
+    }
+  },
+),
+     
         backgroundColor: Colors.white,
         centerTitle: true,
         automaticallyImplyLeading: false,
         title: const Text(
-          "Wellcome",
+          "Welcome",
           style: TextStyle(
               color: Color(0xff2476BD),
               fontSize: 20), // Optional: to change the text color
